excel_etl_pipeline/claude_requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
psycopg2-binary==2.9.9
pyyaml==6.0.1
pandas==2.1.0
pyarrow==14.0.1
tabulate==0.9.0
requests==2.31.0
